requirements.txt
# Core dependencies - Updated for new LangChain
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.0.2
langchain-community>=0.0.10
langchain-core>=0.1.0
openai>=1.6.1

# Vector database
chromadb>=0.4.0

# Document processing
pypdf2>=3.0.0
python-docx>=1.0.0
markdown>=3.4.0
beautifulsoup4>=4.12.0

# Text processing
sentence-transformers>=2.2.0
tiktoken>=0.5.0

# Reranking
rank-bm25>=0.2.0

# Utilities
python-dotenv>=1.0.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.15.0
requests>=2.28.0